+++ src/rest/rest.application.ts
@@ -5,8 +5,7 @@ import {Component} from '../shared/types/index.js';
 import {DatabaseClient} from '../shared/libs/database-client/index.js';
 import {getMongoURI} from '../shared/helpers/index.js';
 import express, {Express} from 'express';
-import {Controller, ExceptionFilter} from '../shared/libs/rest/index.js';
-import {ParseTokenMiddleware} from '../shared/libs/rest/middleware/parse-token.middleware.js';
+import {Controller, ExceptionFilter, ParseTokenMiddleware} from '../shared/libs/rest/index.js';
 
 
 @injectable()
@@ -22,6 +21,8 @@ export class RestApplication {
     @inject(Component.CommentController) private readonly commentController: Controller,
     @inject(Component.ExceptionFilter) private readonly appExceptionFilter: ExceptionFilter,
     @inject(Component.AuthExceptionFilter) private readonly authExceptionFilter: ExceptionFilter,
+    @inject(Component.HttpExceptionFilter) private readonly httpExceptionFilter: ExceptionFilter,
+    @inject(Component.ValidationExceptionFilter) private readonly validationExceptionFilter: ExceptionFilter,
   ) {
     this.server = express();
   }
@@ -61,6 +62,8 @@ export class RestApplication {
 
   private async _initExceptionFilters() {
     this.server.use(this.authExceptionFilter.catch.bind(this.authExceptionFilter));
+    this.server.use(this.validationExceptionFilter.catch.bind(this.validationExceptionFilter));
+    this.server.use(this.httpExceptionFilter.catch.bind(this.httpExceptionFilter));
     this.server.use(this.appExceptionFilter.catch.bind(this.appExceptionFilter));
   }
