+++ src/rest/rest.container.ts
@@ -4,7 +4,8 @@ import {Component} from '../shared/types/index.js';
 import {Config, RestConfig, RestSchema} from '../shared/libs/config/index.js';
 import {Container} from 'inversify';
 import {Logger, PinoLogger} from '../shared/libs/logger/index.js';
-import {AppExceptionFilter, ExceptionFilter} from '../shared/libs/rest/index.js';
+import {AppExceptionFilter, ExceptionFilter, ValidationExceptionFilter} from '../shared/libs/rest/index.js';
+import {HttpErrorExceptionFilter} from '../shared/libs/rest/exception-filter/http-error.exception-filter.js';
 
 export function createRestApplicationContainer() {
   const restApplicationContainer = new Container();
@@ -14,6 +15,7 @@ export function createRestApplicationContainer() {
   restApplicationContainer.bind<Config<RestSchema>>(Component.Config).to(RestConfig).inSingletonScope();
   restApplicationContainer.bind<DatabaseClient>(Component.DatabaseClient).to(MongoDatabaseClient).inSingletonScope();
   restApplicationContainer.bind<ExceptionFilter>(Component.ExceptionFilter).to(AppExceptionFilter).inSingletonScope();
-
+  restApplicationContainer.bind<ExceptionFilter>(Component.HttpExceptionFilter).to(HttpErrorExceptionFilter).inSingletonScope();
+  restApplicationContainer.bind<ExceptionFilter>(Component.ValidationExceptionFilter).to(ValidationExceptionFilter).inSingletonScope();
   return restApplicationContainer;
 }
