+++ src/shared/libs/rest/index.ts
@@ -1,7 +1,8 @@
 export * from './controller/controller.interface.js';
 export * from './controller/base-controller.abstract.js';
 export * from './exception-filter/exception-filter.interface.js';
-export * from './exception-filter/app-exception-filter.js';
+export * from './exception-filter/app.exception-filter.js';
+export * from './exception-filter/validation.exception-filter.js';
 export * from './types/http-method.enum.js';
 export * from './types/route.interface.js';
 export * from './types/request-body.type.js';
